StatGeochemBase

Documentation for StatGeochemBase.

StatGeochemBase.cntr

StatGeochemBase.count_unique!

StatGeochemBase.draw_from_distribution

StatGeochemBase.draw_from_distribution!

StatGeochemBase.findclosest

StatGeochemBase.findclosestabove

StatGeochemBase.findclosestbelow

StatGeochemBase.findmatches

StatGeochemBase.findnth

StatGeochemBase.midpointquadrature

StatGeochemBase.trapezoidalquadrature

StatGeochemBase.cntr — Method

cntr(edges::AbstractArray)

Given an array of bin edges, return a corresponding vector of bin centers

StatGeochemBase.count_unique! — Method

n = count_unique!(A)

Sort the array A in-place (if not already sorted), move unique elements to the front, and return the number of unique elements found.

A[1:count_unique!(A)] should return an array equivalent to unique(A)

StatGeochemBase.draw_from_distribution! — Method

draw_from_distribution!(dist::AbstractArray{<:AbstractFloat}, x::Array{<:AbstractFloat})

Fill an existing variable x with random floating point numbers drawn from a continuous probability distribution specified by a vector dist defining the PDF curve thereof.

StatGeochemBase.draw_from_distribution — Method

x = draw_from_distribution(dist::AbstractArray{<:AbstractFloat}, n::Integer)

Draw n random floating point numbers from a continuous probability distribution specified by a vector dist defining the PDF curve thereof.

StatGeochemBase.findclosest — Method

findclosest(source, target)

Return the index of the numerically closest value in the indexable collection target for each value in source. If muliple values are equally close, the first one is used

StatGeochemBase.findclosestabove — Method

findclosestabove(source, target)

Return the index of the nearest value of the indexable collection target that is greater than (i.e., "above") each value in source. If no such values exist in target, returns an index of 0.

StatGeochemBase.findclosestbelow — Method

findclosestbelow(source, target)

Return the index of the nearest value of the indexable collection target that is less than (i.e., "below") each value in source. If no such target values exist in target, returns an index of 0.

StatGeochemBase.findmatches — Method

findmatches(source, target)

Return the index of the first value in target (if any) that is equal to a given value in source for each value in source; else 0.

StatGeochemBase.findnth — Method

findnth(t::AbstractArray{Bool}, n::Integer)

Return the index of the nth true value of t, else length(t)

StatGeochemBase.midpointquadrature — Method

midpointquadrature(bincenters, values)

Add up the area under a curve with y positions specified by a vector of values and x positions specfied by a vector of bincenters using midpoint integration.

StatGeochemBase.trapezoidalquadrature — Method

trapezoidalquadrature(edges, values)

Add up the area under a curve with y positions specified by a vector of values and x positions specfied by a vector of edges using trapezoidal integration. Bins need not be evenly spaced, though it helps (integration will be faster if edges are specified as an AbstractRange).