StatGeochemBase

Documentation for StatGeochemBase.

StatGeochemBase.arcdistance

StatGeochemBase.cntr

StatGeochemBase.containsi

StatGeochemBase.count_unique!

StatGeochemBase.draw_from_distribution

StatGeochemBase.draw_from_distribution!

StatGeochemBase.fast_inv_sqrt

StatGeochemBase.find_grid_inpolygon

StatGeochemBase.findclosest

StatGeochemBase.findclosestabove

StatGeochemBase.findclosestbelow

StatGeochemBase.findclosestunequal

StatGeochemBase.findmatches

StatGeochemBase.findnth

StatGeochemBase.imsc

StatGeochemBase.imsci

StatGeochemBase.inpolygon

StatGeochemBase.leastsigfig

StatGeochemBase.linreg

StatGeochemBase.linterp1

StatGeochemBase.linterp1!

StatGeochemBase.linterp1s

StatGeochemBase.linterp1s!

StatGeochemBase.midpointquadrature

StatGeochemBase.minarcdistance

StatGeochemBase.nearest

StatGeochemBase.norm_quantile

StatGeochemBase.norm_width

StatGeochemBase.normcdf

StatGeochemBase.normcdf!

StatGeochemBase.normcdf_ll

StatGeochemBase.normcdf_ll!

StatGeochemBase.normpdf

StatGeochemBase.normpdf_ll

StatGeochemBase.normproduct

StatGeochemBase.normproduct_ll

StatGeochemBase.rescale

StatGeochemBase.sigdigits

StatGeochemBase.trapezoidalquadrature

StatGeochemBase.arcdistance — Method

arcdistance(latᵢ,lonᵢ,lat,lon)

Calculate the distance on a sphere between the point (latᵢ,lonᵢ) and any number of points in (lat,lon). Latitude and Longitude should be specified in decimal degrees

StatGeochemBase.cntr — Method

cntr(edges::Collection)

Given an array of bin edges, return a corresponding vector of bin centers

Examples

julia> cntr(1:10)
 1.5:1.0:9.5

StatGeochemBase.containsi — Method

containsi(haystack, needle)

Converts both haystack and needle to strings and checks whether string(haystack) contains string(needle), ignoring case.

StatGeochemBase.count_unique! — Method

n = count_unique!(A)

Sort the array A in-place (if not already sorted), move unique elements to the front, and return the number of unique elements found.

A[1:count_unique!(A)] should return an array equivalent to unique(A)

Examples

julia> A = rand(1:5, 10)
10-element Vector{Int64}:
 4
 4
 2
 3
 3
 4
 1
 5
 1
 2

julia> A = rand(1:5, 7)
7-element Vector{Int64}:
 1
 1
 4
 3
 1
 1
 4

julia> n = count_unique!(A)
3

julia> A
7-element Vector{Int64}:
 1
 3
 4
 1
 3
 4
 4

julia> A[1:n]
3-element Vector{Int64}:
 1
 3
 4

StatGeochemBase.draw_from_distribution! — Method

draw_from_distribution!(x::DenseArray{<:AbstractFloat}, dist::Collection{AbstractFloat})

Fill an existing variable x with random floating point numbers drawn from a continuous probability distribution specified by a vector dist defining the PDF curve thereof.

StatGeochemBase.draw_from_distribution — Method

x = draw_from_distribution(dist::Collection{AbstractFloat}, n::Integer)

Draw n random floating point numbers from a continuous probability distribution specified by a collection dist defining the PDF curve thereof.

StatGeochemBase.fast_inv_sqrt — Method

fast_inv_sqrt(x)

The infamous fast inverse square root of x, in 32 and 64 bit versions. Can be up to 10x faster than base 1/sqrt(x), though with nontrivial loss of precision. The implementations here are good to about 4 ppm.

StatGeochemBase.find_grid_inpolygon — Method

(columns, rows) = find_grid_inpolygon(grid_x, grid_y, poly_x, poly_y)

Find the indexes of grid points that fall within a polygon for a grid with cell centers given by gridx (j-columns of grid) and gridy (i-rows of grid). Returns a list of rows and columns in the polygon

Examples

```julia julia> grid_x = -1.5:1/3:1.5;

julia> grid_y = -1.5:1/3:1.5;

julia> cols,rows = findgridinpolygon(gridx, gridy, [-.4,.4,.4,-.4],[.4,.4,-.4,-.4]) ([5, 5, 6, 6], [5, 6, 5, 6])

julia> grid_x[cols] 4-element Vector{Float64}: -0.16666666666666666 -0.16666666666666666 0.16666666666666666 0.16666666666666666

julia> grid_y[rows] 4-element Vector{Float64}: -0.16666666666666666 0.16666666666666666 -0.16666666666666666 0.16666666666666666

StatGeochemBase.findclosest — Method

findclosest(source, target)

Return the index of the numerically closest value in the indexable collection target for each value in source. If muliple values are equally close, the first one is used

Examples

StatGeochemBase.findclosestabove — Method

findclosestabove(source, target)

Return the index of the nearest value of the indexable collection target that is greater than (i.e., "above") each value in source. If no such values exist in target, returns an index of 0.

StatGeochemBase.findclosestbelow — Method

findclosestbelow(source, target)

Return the index of the nearest value of the indexable collection target that is less than (i.e., "below") each value in source. If no such target values exist in target, returns an index of 0.

StatGeochemBase.findclosestunequal — Method

findclosestunequal(x::Collection, i::Integer)

Return the index of the closest index n to i for which x[n] != x[i], or i if no unequal values of x are found.

StatGeochemBase.findmatches — Method

findmatches(source, target)

Return the linear index of the first value in target (if any) that is equal to a given value in source for each value in source; else 0.

Examples

julia> findmatches([3,5],1:10)
2-element Vector{Int64}:
 3
 5

StatGeochemBase.findnth — Method

findnth(t::Collection{Bool}, n::Integer)

Return the index of the nth true value of t, else length(t)

StatGeochemBase.imsc — Function

imsc(A::AbstractArray, colormap::AbstractVector=viridis, cmin=nanminimum(A), cmax=nanmaximum(A))

Convert a matrix A to an image (an array of Colors.jl colors) using the specified colormap (default viridis), optionally scaled between cmin and cmax.

Examples

julia> A = rand(3,3)
3×3 Matrix{Float64}:
 0.39147   0.553489  0.351628
 0.331786  0.343836  0.824674
 0.639233  0.558113  0.965627

julia> img = imsc(A) # N.B. will display as image if `using ImageInTerminal`
3×3 Array{RGB{N0f8},2} with eltype ColorTypes.RGB{FixedPointNumbers.N0f8}:
 RGB{N0f8}(0.282,0.137,0.455)  …  RGB{N0f8}(0.278,0.051,0.376)
 RGB{N0f8}(0.267,0.004,0.329)     RGB{N0f8}(0.431,0.808,0.345)
 RGB{N0f8}(0.133,0.553,0.553)     RGB{N0f8}(0.992,0.906,0.145)

julia> using Images; save("img.png", img) # Save to file as PNG

julia> using Plots; plot(0:3, 0:3, img) # Plot with specified x and y axes

StatGeochemBase.imsci — Function

imsci(A::AbstractArray, colormap::AbstractVector=viridis, cmin=nanminimum(A), cmax=nanmaximum(A))

Convert a matrix A to an indirect array image (an IndirectArray of Colors.jl colors) using the specified colormap (default viridis), optionally scaled between cmin and cmax.

As imsc, but returns an IndirectArray; slightly more space efficient for small colormaps, but with computational cost.

Examples

julia> A = rand(3,3)
3×3 Matrix{Float64}:
 0.39147   0.553489  0.351628
 0.331786  0.343836  0.824674
 0.639233  0.558113  0.965627

 julia> img = imsci(A)
 3×3 IndirectArrays.IndirectArray{RGB{N0f8}, 2, UInt64, Matrix{UInt64}, Vector{RGB{N0f8}}}:
  RGB{N0f8}(0.282,0.137,0.455)  …  RGB{N0f8}(0.278,0.051,0.376)
  RGB{N0f8}(0.267,0.004,0.329)     RGB{N0f8}(0.431,0.808,0.345)
  RGB{N0f8}(0.133,0.553,0.553)     RGB{N0f8}(0.992,0.906,0.145)

julia> using Images; save("img.png", img) # Save to file as PNG

julia> using Plots; plot(0:3, 0:3, img) # Plot with specified x and y axes

StatGeochemBase.inpolygon — Method

inpolygon(x,y,point)

Check if a 2D polygon defined by the arrays x, y contains a given point. Returns boolean (true or false)

Examples

julia> x = [0, 1, 1, 0];

julia> y = [0, 0, 1, 1];

julia> inpolygon(x, y, (0.5,0.5))
true

julia> inpolygon(x, y, (0.5,1.5))
false

StatGeochemBase.leastsigfig — Method

leastsigfig(d)

Return the order of magnitude of the least significant decimal digit of a number d.

Examples

julia> leastsigfig(1000)
1000.0

julia> leastsigfig(1001)
1.0

julia> leastsigfig(1001.1234)
0.0001

StatGeochemBase.linreg — Method

(a,b) = linreg(x::AbstractVector, y::AbstractVector)

Returns the coefficients for a simple linear least-squares regression of the form y = a + bx

Examples

julia> a, b = linreg(1:10, 1:10)
2-element Vector{Float64}:
 -1.19542133983862e-15
  1.0

julia> isapprox(a, 0, atol = 1e-12)
true

julia> isapprox(b, 1, atol = 1e-12)
true

StatGeochemBase.linterp1! — Method

linterp1!(yq::StridedArray, x::AbstractArray, y::AbstractArray, xq; extrapolate=:Linear, knot_index=ones(Int, length(xq)))

In-place variant of linterp1.

StatGeochemBase.linterp1 — Method

yq = linterp1(x::AbstractArray, y::AbstractArray, xq; extrapolate=:Linear)

Simple linear interpolation in one dimension. Given a vector of knots x and values y, find the corresponding y values at position(s) xq.

Knots x must be sorted in increasing order.

If the optional keyword argument extrapolate is set to :Linear (default), xq values outside the range of x will be extrapolated using a linear extrapolation of the closest two x-y pairs. Otherwise, if extrapolate is set to a Number (e.g., 0, or NaN), that number will be used instead.

Examples

julia> linterp1(1:10, 1:10, 5.5)
5.5

julia> linterp1(1:10, 1:10, 0.5:10.5)
11-element Vector{Float64}:
  0.5
  1.5
  2.5
  3.5
  4.5
  5.5
  6.5
  7.5
  8.5
  9.5
 10.5

StatGeochemBase.linterp1s! — Method

linterp1s!(yq::StridedArray, x::StridedArray, y::StridedArray, xq; extrapolate=:Linear)
linterp1s!(yq::StridedArray, knot_index::StridedArray{Int}, x::StridedArray, y::StridedArray, xq::AbstractArray; extrapolate=:Linear)

In-place variant of linterp1s. Will sort x and permute y to match, before interpolating at xq and storing the result in yq.

An optional temporary working array knot_index = similar(xq, Int) may be provided to fully eliminate allocations.

StatGeochemBase.linterp1s — Method

yq = linterp1s(x::AbstractArray, y::AbstractArray, xq; extrapolate=:Linear)

As as linterp1 (simple linear interpolation in one dimension), but will sort the knots x and values y pairwise if x if not already sorted in increasing order.

Examples

julia> linterp1s(10:-1:1, 1:10, 5.5)
5.5

julia> linterp1s(10:-1:1, 1:10, 0.5:10.5)
11-element Vector{Float64}:
 10.5
  9.5
  8.5
  7.5
  6.5
  5.5
  4.5
  3.5
  2.5
  1.5
  0.5

StatGeochemBase.midpointquadrature — Method

midpointquadrature(bincenters, values)

Add up the area under a curve with y positions specified by a vector of values and x positions specfied by a vector of bincenters using midpoint integration.

Examples

julia> midpointquadrature(0:0.1:10, 0:0.1:10)
50.5

StatGeochemBase.minarcdistance — Method

minarcdistance(latᵢ,lonᵢ,lat,lon)

Return the smallest non-NaN arcdistance (i.e. distance on a sphere in arc degrees) between a given point (latᵢ[i],lonᵢ[i]) and any point in (lat,lon) for each i in eachindex(latᵢ, lonᵢ).

Latitude and Longitude should be specified in decimal degrees

StatGeochemBase.nearest — Method

nearest(T, x)

Convert x to the nearest representable value in type T, rounding if inexact

Examples

```julia julia> nearest(Int, 1234.56) 1235

julia> nearest(Int, Inf) 9223372036854775807

julia> nearest(Int, -Inf) -9223372036854775808 ````

StatGeochemBase.norm_quantile — Method

norm_quantile(F::Number)

How far away from the mean (in units of sigma) should we expect proportion F of the samples to fall in a standard Gaussian (Normal[0,1]) distribution

StatGeochemBase.norm_width — Method

norm_width(N::Number)

How dispersed (in units of sigma) should we expect a sample of N numbers drawn from a standard Gaussian (Normal[0,1]) distribution to be?

StatGeochemBase.normcdf! — Method

normcdf!(result,mu,sigma,x)

In-place version of normcdf

StatGeochemBase.normcdf — Method

normcdf(mu,sigma,x)

Cumulative distribution function of the Normal (Gaussian) distribution

$1/2 + erf( rac{x-μ}{σ√2})/2$

with mean mu and standard deviation sigma, evaluated at x.

StatGeochemBase.normcdf_ll! — Method

normcdf_ll!(mu, sigma, x)

Fast log likelihood corresponding to the natural logarithm of the cumulative distribution function of a Normal (Gaussian) distribution with mean mu and standard deviation sigma, evaluated at x.

As normcdf_ll, but in-place (using x as a buffer).

StatGeochemBase.normcdf_ll — Method

normcdf_ll(mu, sigma, x)

Fast log likelihood corresponding to the natural logarithm of the cumulative distribution function of a Normal (Gaussian) distribution with mean mu and standard deviation sigma, evaluated at x.

If x, [mu, and sigma] are given as arrays, the sum of the log likelihood over all x will be returned.

See also normcdf

StatGeochemBase.normpdf — Method

normpdf(mu,sigma,x)

Probability density function of the Normal (Gaussian) distribution

$ℯ^{-(x-μ)^2 / (2σ^2)} / σ√2π$

with mean mu and standard deviation sigma, evaluated at x

StatGeochemBase.normpdf_ll — Method

normpdf_ll(mu, sigma, x)

Fast log likelihood corresponding to the natural logarithm of the probability density function of a Normal (Gaussian) distribution with mean mu and standard deviation sigma, evaluated at x.

If x, [mu, and sigma] are given as arrays, the sum of the log likelihood over all x will be returned.

See also normpdf

StatGeochemBase.normproduct — Method

normproduct(μ1, σ1, μ2, σ2)

The integral of the product of two normal distributions N[μ1,σ1] * N[μ2,σ2]. This is itself just another Normal distribution! Specifically, one with variance σ1^2 + σ2^2, evaluated at distance |μ1-μ2| from the mean

StatGeochemBase.normproduct_ll — Method

normproduct_ll(μ1, σ1, μ2, σ2)

Log likelihood corresponding to the integral of N[μ1,σ1] * N[μ2,σ2] As normproduct, but using the fast log likelihood of a Normal distribution

StatGeochemBase.rescale — Function

rescale(y, min::Number=0, max::Number=1)

Rescale a collection of numbers y between a new minimum min and new maximum max

Examples

julia> rescale(1:5)
5-element Vector{Float64}:
0.0
0.25
0.5
0.75
1.0

julia> rescale(1:5, -1, 0)
5-element Vector{Float64}:
-1.0
-0.75
-0.5
-0.25
0.0

StatGeochemBase.sigdigits — Method

sigdigits(d)

Determine the number of decimal significant figures of a number d.

Examples

julia> sigdigits(1000)
1

julia> sigdigits(1001)
4

julia> sigdigits(1001.1245)
8

StatGeochemBase.trapezoidalquadrature — Method

trapezoidalquadrature(edges, values)

Add up the area under a curve with y positions specified by a vector of values and x positions specfied by a vector of edges using trapezoidal integration. Bins need not be evenly spaced, though it helps (integration will be faster if edges are specified as an AbstractRange).

Examples

julia> trapezoidalquadrature(0:0.1:10, 0:0.1:10)
50.0