StatGeochemBase

Documentation for StatGeochemBase.

StatGeochemBase.MSWD

StatGeochemBase.arcdistance

StatGeochemBase.awmean

StatGeochemBase.cntr

StatGeochemBase.count_unique!

StatGeochemBase.draw_from_distribution

StatGeochemBase.draw_from_distribution!

StatGeochemBase.fast_inv_sqrt

StatGeochemBase.find_grid_inpolygon

StatGeochemBase.findclosest

StatGeochemBase.findclosestabove

StatGeochemBase.findclosestbelow

StatGeochemBase.findmatches

StatGeochemBase.findnth

StatGeochemBase.gwmean

StatGeochemBase.inpolygon

StatGeochemBase.linreg

StatGeochemBase.midpointquadrature

StatGeochemBase.nearest

StatGeochemBase.norm_quantile

StatGeochemBase.norm_width

StatGeochemBase.normcdf

StatGeochemBase.normcdf!

StatGeochemBase.normpdf

StatGeochemBase.normpdf_ll

StatGeochemBase.normproduct

StatGeochemBase.normproduct_ll

StatGeochemBase.trapezoidalquadrature

StatGeochemBase.yorkfit

StatGeochemBase.MSWD — Method

MSWD(x, σ)

Return the Mean Square of Weighted Deviates (AKA the reduced chi-squared statistic) of a dataset with values x and one-sigma uncertainties σ

StatGeochemBase.arcdistance — Method

arcdistance(latᵢ,lonᵢ,lat,lon)

Calculate the distance on a sphere between the point (latᵢ,lonᵢ) and any number of points in (lat,lon). Latitude and Longitude should be specified in decimal degrees

StatGeochemBase.awmean — Method

(wx, wσ, mswd) = awmean(x, σ)

Weighted mean, absent the geochonologist's MSWD correction to uncertainty.

StatGeochemBase.cntr — Method

cntr(edges::AbstractArray)

Given an array of bin edges, return a corresponding vector of bin centers

StatGeochemBase.count_unique! — Method

n = count_unique!(A)

Sort the array A in-place (if not already sorted), move unique elements to the front, and return the number of unique elements found.

A[1:count_unique!(A)] should return an array equivalent to unique(A)

StatGeochemBase.draw_from_distribution! — Method

draw_from_distribution!(dist::AbstractArray{<:AbstractFloat}, x::Array{<:AbstractFloat})

Fill an existing variable x with random floating point numbers drawn from a continuous probability distribution specified by a vector dist defining the PDF curve thereof.

StatGeochemBase.draw_from_distribution — Method

x = draw_from_distribution(dist::AbstractArray{<:AbstractFloat}, n::Integer)

Draw n random floating point numbers from a continuous probability distribution specified by a vector dist defining the PDF curve thereof.

StatGeochemBase.fast_inv_sqrt — Method

fast_inv_sqrt(x)

The infamous fast inverse square root of x, in 32 and 64 bit versions. Can be up to 10x faster than base 1/sqrt(x), though with nontrivial loss of precision. The implementations here are good to about 4 ppm.

StatGeochemBase.find_grid_inpolygon — Method

(rows, columns) = find_grid_inpolygon(grid_x, grid_y, poly_x, poly_y)

Find the indexes of grid points that fall within a polygon for a grid with cell centers given by gridx (j-columns of grid) and gridy (i-rows of grid). Returns a list of rows and columns in the polygon

StatGeochemBase.findclosest — Method

findclosest(source, target)

Return the index of the numerically closest value in the indexable collection target for each value in source. If muliple values are equally close, the first one is used

StatGeochemBase.findclosestabove — Method

findclosestabove(source, target)

Return the index of the nearest value of the indexable collection target that is greater than (i.e., "above") each value in source. If no such values exist in target, returns an index of 0.

StatGeochemBase.findclosestbelow — Method

findclosestbelow(source, target)

Return the index of the nearest value of the indexable collection target that is less than (i.e., "below") each value in source. If no such target values exist in target, returns an index of 0.

StatGeochemBase.findmatches — Method

findmatches(source, target)

Return the index of the first value in target (if any) that is equal to a given value in source for each value in source; else 0.

StatGeochemBase.findnth — Method

findnth(t::AbstractArray{Bool}, n::Integer)

Return the index of the nth true value of t, else length(t)

StatGeochemBase.gwmean — Method

(wx, wσ, mswd) = gwmean(x, σ)

Geochronologist's weighted mean, with "MSWD correction" to uncertainty, i.e., wσ is increased by a factor of sqrt(mswd)

StatGeochemBase.inpolygon — Method

inpolygon(x,y,point)

Check if a 2D polygon defined by the arrays x, y contains a given point. Returns boolean (true or false)

StatGeochemBase.linreg — Method

(a,b) = linreg(x::AbstractVector, y::AbstractVector)

Returns the coefficients for a simple linear least-squares regression of the form y = a + bx

StatGeochemBase.midpointquadrature — Method

midpointquadrature(bincenters, values)

Add up the area under a curve with y positions specified by a vector of values and x positions specfied by a vector of bincenters using midpoint integration.

StatGeochemBase.nearest — Method

nearest(T, x)

Convert x to the nearest representable value in type T, rounding if inexact

StatGeochemBase.norm_quantile — Method

norm_quantile(F::Number)

How far away from the mean (in units of sigma) should we expect proportion F of the samples to fall in a standard Gaussian (Normal[0,1]) distribution

StatGeochemBase.norm_width — Method

norm_width(N::Number)

How dispersed (in units of sigma) should we expect a sample of N numbers drawn from a standard Gaussian (Normal[0,1]) distribution to be?

StatGeochemBase.normcdf! — Method

normcdf!(result,mu,sigma,x)

In-place version of normcdf

StatGeochemBase.normcdf — Method

normcdf(mu,sigma,x)

Cumulative density function of the Normal (Gaussian) distribution

$1/2 + erf( rac{x-μ}{σ√2})/2$

with mean mu and standard deviation sigma, evaluated at x.

StatGeochemBase.normpdf — Method

normpdf(mu,sigma,x)

Probability density function of the Normal (Gaussian) distribution

$ℯ^{-(x-μ)^2 / (2σ^2)} / σ√2π$

with mean mu and standard deviation sigma, evaluated at x

StatGeochemBase.normpdf_ll — Method

normpdf_ll(mu, sigma, x)

Fast log likelihood corresponding to a Normal (Gaussian) distribution with mean mu and standard deviation sigma, evaluated at x.

If x, [mu, and sigma] are given as arrays, the sum of the log likelihood over all x will be returned.

See also normpdf

StatGeochemBase.normproduct — Method

normproduct(μ1, σ1, μ2, σ2)

The integral of the product of two normal distributions N[μ1,σ1] * N[μ2,σ2]. This is itself just another Normal distribution! Specifically, one with variance σ1^2 + σ2^2, evaluated at distance |μ1-μ2| from the mean

StatGeochemBase.normproduct_ll — Method

normproduct_ll(μ1, σ1, μ2, σ2)

Log likelihood corresponding to the integral of N[μ1,σ1] * N[μ2,σ2] As normproduct, but using the fast log likelihood of a Normal distribution

StatGeochemBase.trapezoidalquadrature — Method

trapezoidalquadrature(edges, values)

Add up the area under a curve with y positions specified by a vector of values and x positions specfied by a vector of edges using trapezoidal integration. Bins need not be evenly spaced, though it helps (integration will be faster if edges are specified as an AbstractRange).

StatGeochemBase.yorkfit — Method

yorkfit(x, σx, y, σy)

Uses the York (1968) two-dimensional least-squares fit to calculate a, b, and uncertanties σa, σb for the equation y = a + bx, given x, y and uncertaintes σx, `σy.